requirements.txt
coreapi==2.3.3
django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
Pillow==9.5.0
